Move group_true_values_into_list to shared module
The function `group_true_values_into_list()` is used for collapsing
multiple values in rows for the same colum and PID into an array. Move
it to a shared module for reuse for other CLI processes that handle
similarly-structured clinical data.
